// mjolksyra-app/generated-client/models/ExerciseResponse.ts
export interface ExerciseResponse {
    /**
     * 
     * @type {string}
     * @memberof ExerciseResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseResponse
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseResponse
     */
    force: string | null;
    /**
     * 
     * @type {string}
     * @memberof ExerciseResponse
     */
    level: string | null;
    /**
     * 
     * @type {string}
     * @memberof ExerciseResponse
     */
    mechanic: string | null;
    /**
     * 
     * @type {string}
     * @memberof ExerciseResponse
     */
    equipment: string | null;
    /**
     * 
     * @type {string}
     * @memberof ExerciseResponse
     */
    category: string | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof ExerciseResponse
     */
    instructions: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof ExerciseResponse
     */
    primaryMuscles: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof ExerciseResponse
     */
    secondaryMuscles: Array<string>;
    /**
     * 
     * @type {boolean}
     * @memberof ExerciseResponse
     */
    starred: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ExerciseResponse
     */
    canDelete: boolean;
}